+++ scripts/add_electricity.py
@@ -1,4 +1,5 @@
 # SPDX-FileCopyrightText: Contributors to PyPSA-Eur <https://github.com/pypsa/pypsa-eur>
+# SPDX-FileCopyrightText: Open Energy Transition gGmbH
 #
 # SPDX-License-Identifier: MIT
 
@@ -405,6 +406,7 @@ def set_transmission_costs(
 def attach_wind_and_solar(
     n: pypsa.Network,
     costs: pd.DataFrame,
+    ppl: pd.DataFrame,
     profile_filenames: dict,
     carriers: list | set,
     extendable_carriers: list | set,
@@ -420,6 +422,8 @@ def attach_wind_and_solar(
         The PyPSA network to attach the generators to.
     costs : pd.DataFrame
         DataFrame containing the cost data.
+    ppl : pd.DataFrame
+        DataFrame containing the power plant data.
     profile_filenames : dict
         Dictionary containing the paths to the wind and solar profiles.
     carriers : list | set
@@ -493,12 +497,21 @@ def attach_wind_and_solar(
             p_max_pu = ds["profile"].to_pandas()
             p_max_pu.columns = p_max_pu.columns.map(flatten)
 
+            if not ppl.query("carrier == @car").empty:
+                caps = ppl.query("carrier == @car").groupby("bus").p_nom.sum()
+                caps = pd.Series(data=caps, index=ds.indexes["bus"]).fillna(0)
+            else:
+                caps = pd.Series(index=ds.indexes["bus"]).fillna(0)
+            caps.index = caps.index.map(flatten)
+
             n.add(
                 "Generator",
                 bus_bins,
                 suffix=" " + car,
                 bus=buses,
                 carrier=car,
+                p_nom=caps,
+                p_nom_min=caps,
                 p_nom_extendable=car in extendable_carriers["Generator"],
                 p_nom_max=p_nom_max,
                 marginal_cost=costs.at[supcar, "marginal_cost"],
@@ -1123,6 +1136,7 @@ if __name__ == "__main__":
     attach_wind_and_solar(
         n,
         costs,
+        ppl,
         snakemake.input,
         renewable_carriers,
         extendable_carriers,